Change locationareadetails struct add commandExplore, locationEncounters
Changed the LocationAreaDetails struct to not be inline so its internal
structs can be accessed outside.

Added the fully implemented commandExplore funcion which gets the
pokemon encounters in the location and prints them out.

Added a stub for locationEncounters which will make a GET request to the
pokeAPI to get location-area details. It will then return a slice of
PokemonEncounters.